new version
git-svn-id: https://plugins.svn.wordpress.org/easy-code-placement/trunk@923340 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/ecp_delete.php
@@ -2,9 +2,23 @@
 
 global $wpdb;
 
-$ecp_id = $_GET['ecpid'];
+// secure get data and set variables
+$_GET = stripslashes_deep($_GET);
+$ecp_id = intval($_GET['ecpid']);
+
+if($ecp_id=="" || !is_numeric($ecp_id)) {
+  // when get emty or other than numbers goto error page
+	$ecp_error = __('Modifying of the ID is not allowed', 'ecp');
+  $ecp_error_page = "";
+  $ecp_error_id = "";
+  ecp_error($ecp_error, $ecp_error_page, $ecp_error_id);
+  exit();
+}
+
+// delete code
 $wpdb->delete($wpdb->prefix.'ecp_data', array('id'=>$ecp_id));
-	
+
+// when data is deleted goto options page	
 header('Location: options-general.php?page=ecp_option_page');
 
 ?>
